backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21          # SQLModel = SQLAlchemy + Pydantic in one
aiosqlite==0.20.0
apscheduler==3.10.4       # Background scheduler for monitor checks
httpx==0.27.2             # Async HTTP for health checks + UniFi API
pyyaml==6.0.2
python-multipart==0.0.9
python-nmap==0.7.1        # nmap wrapper for device discovery
mac-vendor-lookup==0.1.12 # MAC OUI → vendor name
icmplib==3.0.4            # Pure-Python ICMP ping (no subprocess)
